Prevent errors with empty QuickFix buffer
Currently quickr-preview will spit the dummy when pressing 'enter' or
'<leader>+space' within a empty QuickFix buffer. This is occurring as
quickr-preview is attempting to index s:qflist, despite the fact that
its empty. This patch resolves the issue by fist checking if s:qflist
is empty before attempting to index it; if so then we have nothing to
do as the the QuickFix buffer must be empty.