Prevent QuickFix and Location Lists from stomping on each other
The qf list and qf len are currently defines as script variables. This
works fine until the QuickFix and Location List are open at the same
time, at which point they start fighting over the same instances of the
two variables. This patch resolves the issue by defining qf list and qf
len as buffer variables, allowing the QuickFix and Location List to
manage their own instances of the two variables.