Prevent Location List from grabbing QuickFix list
Currently GenerateBufferList checks first for an existing QuickFix list
and, if finds none, then for an existing Location list. The problem is
that the QuickFix list is global to all windows/buffers. This means that
opening a Location window when a QuickFix list already exists will cause
the QuickFix list (not the Location list) to be assigned to the Location
window. This patch resolves this issue be swapping the ordering that the
GenerateBufferList checks for QuickFix/Location lists. This fix works as
Location lists are local to the current window/buffer, therefore should
never exist for a QuickFix window.