fix bug: buffer settings not cleared when re-opening buffer
Sometimes I find that the buffer's local settings are not being cleared
when opening a buffer from the qf/loc list. This can be frustrating, as
this results in the buffer remaining with no swap file, etc.

The solution is to delete the buffer before re-opening it outside of the
preview window. Doing this appears to flush out any settings that were
previously set by 'setlocal' when the buffer was opened in the preview
window.