fix bug: qf/loc window grows in size when closing preview window
This seems to be a common bug:

https://github.com/mileszs/ack.vim/issues/150
https://github.com/vim-syntastic/syntastic/issues/1489

For me, the issue occurs when I have NERDTree open; each time I close
the preview window, the qf/loc window grows to fill the space that was
previously filled by the preview window. This will continue to occur on
each preview window close, until the qf/loc window takes up the entire
screen.

The fix introduced by this patch is to force the qf/loc window back to
its original size each time the preview window closed. This fix is a bit
hacky, but it seems to work well.