fix bug: preview window size incorrect when vertical splits present
When opening the preview window above the qf/loc list, it will expand to
fill half of the screen realestate available to vim. This is ideal as it
gives a nice balance between the buffer being previewed, and the buffer
being edited. However whenever a vertical split is present (for example
NerdTree or TagBar) things fall apart. In this case the preview window
only expands to half that of the qf/loc window. This results in it being
only a few lines in size, which is most definitely less than ideal.

This patch resolves this issue by switching back to `pedit` and manually
calculating/forcing the preview window height/width. I've tried several
different solutions, but this is the one the works in all cases, that is
when opening the preview above/below/left/right the qf/loc window.