bug fix: highlighting persists when closing preview window
The current line lighting persists when closing the preview window. This
is not usually an issue, unless the same buffer exists in another window.
This patch deletes the current line highlighting before when closing the
preview window, therefore also deleting it from any other windows which
may contain the same buffer.