enhancement: improve performance when previewing opened buffers
I often find myself poking around very large files (100k+ lines), and
this is where performance takes a hit. This patch enhances performance
by removing unnecessary calls to `pedit` when the buffer of interest is
already opened in the preview window. The performance increase is quite
noticeable (as least with the code base that I am working on), and also
eliminates flickering cause by the closing and reopening of the preview
window.