fix bug: plugin loses track of "already listed" buffers
When a qf/loc buffer is created, this plugin creates a list of already
opened buffers. This list is used to ensure the already opened buffers
are not set to unlisted/read-only/no-swap when opened in the preview
window.

The problem is this list is not updated when new buffers are opened or
when existing buffers are deleted. Therefore the list no longer tracks
the current state of listed buffers, and the wrong settings may be set
when these buffers opened in the preview window.

This patch resolves this issue by removing the already opened buffers
list completely. It is replaced by simply querying whether the buffer
is listed just before opening it in the preview window. This seems to
work well, and reduces the complexity of the code.