fix bug: plugin allows invalid buffers to be opened
When pressing `enter` within the qf/loc window, the plugin will attempt
to open the qf/loc entry, regardless of whether its referencing a valid
or invalid buffer. This is not ideal, as opening invalid buffers causes
weird behaviors. This problem occurs due to the simulated `enter` being
invoked outside the HandleEnterQuickfix function.

This patch resolves the issue by moving the simulated 'enter' key press
to inside the HandleEnterQuickfix function; this means that it will only
be invoked when the error checking (GetValidEntry function) is successful.