enhancement: improved cleanup when buffer leaves preview window
This patch simplifies cleanup (reverting unlisted/read-only/no-swap
buffer settings) when a buffer leaves the preview window. This has
been achieved by settings the `bufhidden` option to `delete`. This
means the buffer will be deleted when it leaves the preview window
for any reason; causing the buffer's local variable to be reset.

It's worth noting that this cleanup is triggered whenever buffers
leave the preview for any reason (not only when manually triggered
by this plugin).