fix bug: highlighting applied outside of preview window
The "current result" highlighting is applied to all windows containing
the buffer of interest, not just the preview window. This can be shown
when previewing qf/loc results for a buffer that is already displayed.

This patch resolves this issue by using `match` instead of `sign` for
highlighting the result; as it applies to the local window only. This
also means that the highlighting does not need to be cleaned up after
the preview window is closed. A further benefit of `match` is that it
simplifies this plugins configuration as it doesn't require the usage
of the sign column.