fix bug: bufdelete auto cmd lost when opening second qf/loc list
The `BufDelete` auto command (assigned to the qf/loc buffer) is deleted
when another qf/loc buffer is opened. This occurs because `autocmd!` is
applied to all buffers, not just the new buffer being created.

This patch resolves the issue by ensuring that only auto commands which
are assigned to the current buffer are deleted. This has been achieved
with the usage of `autocmd! * <buffer>`.