new feature: option to select preview window settings
This new option gives the user the ability to define what setting should
be applied to a buffer when it is opened in the preview window. This is
useful for disabling additional settings such as signcolumn, colorcolumn,
cursorcolumn, cursorline, etc. If this setting is not specified, it will
default to:

let g:quickr_preview_options = 'number norelativenumber nofoldenable'