enhancement: update local copy of qf/loc list when it has changed
This plugin takes a local copy of the qf/loc list (`b:qflist`) when the
qf/loc buffer is first created. This is useful, as using the local copy
is a lot faster than invoking `getqflist()`/`getloclist()` each time we
want to preview or enter a buffer. This is most noticeable when working
with large qf/loc lists.

The problem is that this locally cached qf/loc list is not updated when
the underlying qf/loc list changes. This patch resolves this problem by
detecting if the underlying qf/loc list changes, and if so, updates the
locally cached qf/loc list (`b:qflist`).

The trick is NOT to request the qf/loc list when determining whether it
has changed, as it renders the entire process moot. Fortunately Vim 8.1
provides the `quickfix-changedtick`, a variable that's incremented each
time the qf/loc list is changed. Most importantly, this variable can be
accessed without requesting the entire qf/loc list.

It is important to note that the patch only fixes the issue for Vim 8.1
(or greater). For previous Vim versions the behavior has not changed.