fix bug: folds remain closed when entering a buffer
When pressing `enter` to open a buffer from the qf/loc list, the plugin
will fail to jump to the correct line number when that line is currently
hidden within a fold.

This patch resolves the issue by placing a `foldopen` directly after the
simulated `enter` is invoked. Note that the `silent!` is there to shutup
the error message that occurs when no fold is present.